Refactor tests and css files
